+++ R/read_shimadzu_lcd.R
@@ -1,5 +1,71 @@
 #' Shimadzu LCD parser
 #'
+#' Read 3D PDA or 2D chromatogram streams from 'Shimadzu' LCD files.
+#'
+#' A parser to read data from 'Shimadzu' \code{.lcd} files. LCD files are
+#' encoded as 'Microsoft' OLE documents. The parser relies on the
+#' [olefile](https://pypi.org/project/olefile/) package in Python to unpack the
+#' files. The PDA data is encoded in a stream called \code{PDA 3D Raw Data:3D Raw Data}.
+#' The PDA data stream contains a segment for each retention time, beginning
+#' with a 24-byte header.
+#'
+#' The 24 byte header consists of the following fields:
+#' * 4 bytes: segment label (\code{17234}).
+#' * 4 bytes: ???
+#' * 4 bytes: Little-endian integer specifying the number of wavelength values
+#' in the segment.
+#' * 4 bytes: Little-endian integer specifying the total number of bytes in the segment.
+#' * 8 bytes of \code{00}s
+#'
+#' Each segment is divided into two sub-segments, which begin and end with an
+#' integer specifying the length of the sub-segment in bytes. All known values
+#' in this data stream are little-endian and the data are delta-encoded. The
+#' first hexadecimal digit of each value is a sign digit
+#' specifying the number of bytes in the delta and whether the value is positive
+#' or negative. The sign digit represents the number of hexadecimal digits used
+#' to encode each value. Even numbered sign digits correspond to positive deltas,
+#' whereas odd numbers indicate negative deltas. Positive values are encoded as
+#' little-endian integers, while negative values are encoded as two's
+#' complements. The value at each position is derived by subtracting the delta
+#' from the previous value.
+#'
+#' @param path Path to LCD file.
+#' @param what What stream to get. Either \code{pda} or \code{chromatogram}. If
+#' stream is not specified, will default to \code{pda} if PDA stream is present.
+#' @param format_out Matrix or data.frame.
+#' @param data_format Either \code{wide} (default) or \code{long}.
+#' @param read_metadata Logical. Whether to attach metadata.
+#' @author Ethan Bass
+#' @return A 3D chromatogram from the PDA stream in \code{matrix} or
+#' \code{data.frame} format, according to the value of \code{format_out}.
+#' The chromatograms will be returned in \code{wide} or \code{long} format
+#' according to the value of \code{data_format}.
+#' @note This parser is experimental and may still need some work. It is not
+#' yet able to interpret much metadata from the files.
+#' @export
+
+read_shimadzu_lcd <- function(path, what, format_out = c("matrix", "data.frame"),
+                                 data_format = c("wide", "long"),
+                                 read_metadata = TRUE){
+  format_out <- match.arg(format_out, c("matrix", "data.frame"))
+  data_format <- match.arg(data_format, c("wide", "long"))
+  if (missing(what)){
+    what <- ifelse(check_streams(path, "pda", boolean=TRUE), "pda", "chromatogram")
+  }
+  olefile_installed <- reticulate::py_module_available("olefile")
+  if (!olefile_installed){
+    configure_python_environment(parser = "olefile")
+  }
+
+  read_sz <- switch(what, "pda" = read_shimadzu_lcd_3D,
+                "chromatogram" = read_shimadzu_lcd_2D)
+
+  read_sz(path, format_out = format_out, data_format = data_format,
+                         read_metadata = read_metadata)
+}
+
+#' Shimadzu LCD 3D parser
+#'
 #' Read 3D PDA data stream from 'Shimadzu' LCD files.
 #'
 #' A parser to read PDA data from 'Shimadzu' \code{.lcd} files. LCD files are
@@ -38,27 +104,21 @@
 #' \code{data.frame} format, according to the value of \code{format_out}.
 #' The chromatograms will be returned in \code{wide} or \code{long} format
 #' according to the value of \code{data_format}.
-#' @note This parser is experimental and may still
-#' need some work. It is not yet able to interpret much metadata from the files.
+#' @note This parser is experimental and may still need some work. It is not
+#' yet able to interpret much metadata from the files.
 #' @export
 
-read_shimadzu_lcd <- function(path, format_out = c("matrix", "data.frame"),
+read_shimadzu_lcd_3D <- function(path, format_out = c("matrix", "data.frame"),
                               data_format = c("wide", "long"),
                               read_metadata = TRUE){
-  format_out <- match.arg(format_out, c("matrix", "data.frame"))
-  data_format <- match.arg(data_format, c("wide", "long"))
-
-  olefile_installed <- reticulate::py_module_available("olefile")
-  if (!olefile_installed){
-    configure_python_environment(parser = "olefile")
-  }
+  check_streams(path, what = "pda")
 
   # read wavelengths from "Wavelength Table" stream
   lambdas <- read_shimadzu_wavelengths(path)
   n_lambdas <- length(lambdas)
 
   # read data from "3D Raw Data" stream
-  dat <- read_shimadzu_raw(path, n_lambdas = n_lambdas)
+  dat <- read_shimadzu_pda(path, n_lambdas = n_lambdas)
   colnames(dat) <- lambdas
 
   # infer times from "PDA.1.Method" stream
@@ -78,14 +138,104 @@ read_shimadzu_lcd <- function(path, format_out = c("matrix", "data.frame"),
   dat
 }
 
+#' Shimadzu LCD 2D parser
+#'
+#' Read 2D PDA data stream from 'Shimadzu' LCD files.
+#'
+#' A parser to read chromatogram data streams from 'Shimadzu' \code{.lcd} files.
+#' LCD files are encoded as 'Microsoft' OLE documents. The parser relies on the
+#' [olefile](https://pypi.org/project/olefile/) package in Python to unpack the
+#' files. The PDA data is encoded in a stream called \code{PDA 3D Raw Data:3D Raw Data}.
+#' The PDA data stream contains a segment for each retention time, beginning
+#' with a 24-byte header.
+#'
+#' The 24 byte header consists of the following fields:
+#' * 4 bytes: segment label (\code{17234}).
+#' * 4 bytes: ???
+#' * 4 bytes: Little-endian integer specifying the number of wavelength values
+#' in the segment.
+#' * 4 bytes: Little-endian integer specifying the total number of bytes in the segment.
+#' * 8 bytes of \code{00}s
+#'
+#' Each segment is divided into multiple sub-segments, which begin and end with an
+#' integer specifying the length of the sub-segment in bytes. All known values
+#' in this data stream are little-endian and the data are delta-encoded. The
+#' first hexadecimal digit of each value is a sign digit
+#' specifying the number of bytes in the delta and whether the value is positive
+#' or negative. The sign digit represents the number of hexadecimal digits used
+#' to encode each value. Even numbered sign digits correspond to positive deltas,
+#' whereas odd numbers indicate negative deltas. Positive values are encoded as
+#' little-endian integers, while negative values are encoded as two's
+#' complements. The value at each position is derived by subtracting the delta
+#' from the previous value. Some chromatograms must be multiplied by a scaling
+#' factor which I have not yet been able to locate in the file.
+#'
+#' @param path Path to LCD file.
+#' @param format_out Matrix or data.frame.
+#' @param data_format Either \code{wide} (default) or \code{long}.
+#' @param read_metadata Logical. Whether to attach metadata.
+#' @author Ethan Bass
+#' @return A 2D chromatogram from the PDA stream in \code{matrix} or
+#' \code{data.frame} format, according to the value of \code{format_out}.
+#' The chromatograms will be returned in \code{wide} or \code{long} format
+#' according to the value of \code{data_format}.
+#' @note This parser is experimental and may still need some work. It is not
+#' yet able to interpret much metadata from the files.
+#' @export
+
+read_shimadzu_lcd_2D <- function(path, format_out = c("matrix", "data.frame"),
+                                 data_format = c("wide", "long"),
+                                 read_metadata = TRUE){
+
+  existing_streams <- check_streams(path, what="chromatogram")
+
+  dat <- lapply(existing_streams, function(stream){
+    read_shimadzu_chrom(path, stream = stream)
+  })
+
+  # infer times from "PDA.1.Method" stream
+  method_metadata <- read_sz_method(path, stream = c("GUMM_Information", "ShimadzuLC.1",
+                                                     "GUC.1.METHOD"))
+  times <- get_sz_times(method_metadata, what="chromatogram", nval = nrow(dat[[1]]))
+
+  if (inherits(times, "numeric")){
+    if (data_format == "wide"){
+      dat <- lapply(dat, function(x){
+        rownames(x) <- times
+        x
+      })
+    } else if (data_format == "long"){
+      dat <- lapply(dat, function(x){
+        cbind(rt=times,x)
+      })
+    }
+  }
+
+  if (format_out == "matrix"){
+    dat <- lapply(dat, function(x){
+      as.matrix(x)
+    })
+  }
+  if (length(dat) == 1){
+    dat <- dat[[1]]
+  }
+  dat
+}
+
+read_shimadzu_chrom <- function(path, stream){
+  path_raw <- export_stream(path, stream = stream)
+  f <- file(path_raw, "rb")
+  on.exit(close(f))
+  data.frame(int = decode_shimadzu_block(f))
+}
+
 #' Read Shimadzu "Method" stream
 #' This function is called internally by \code{read_shimadzu_lcd}.
 #' @author Ethan Bass
 #' @noRd
-read_sz_method <- function(path){
-  method_path <- export_stream(path,
-                                   stream = c("GUMM_Information", "ShimadzuPDA.1",
-                                              "PDA.1.METHOD"),
+read_sz_method <- function(path, stream = c("GUMM_Information", "ShimadzuPDA.1",
+                                            "PDA.1.METHOD")){
+  method_path <- export_stream(path, stream = stream,
                                    remove_null_bytes = TRUE)
   if (is.na(method_path)){
     warning("Method stream could not be found -- unable to infer retention times.")
@@ -118,12 +268,15 @@ read_sz_method <- function(path){
 #' This function is called internally by \code{read_shimadzu_lcd}.
 #' @author Ethan Bass
 #' @noRd
-get_sz_times <- function(sz_method, nval){
-    start_time <- try(get_metadata_field(sz_method, "StTm")/60000, silent = TRUE)
-    end_time <- try(get_metadata_field(sz_method, "EdTm")/60000, silent = TRUE)
-    if (inherits(start_time, "numeric") & inherits(end_time, "numeric")){
-      seq(from = start_time, to = end_time, length.out = nval)
-    } else NA
+get_sz_times <- function(sz_method, what = c("pda","chromatogram"), nval){
+  what <- match.arg(what, c("pda","chromatogram"))
+  fields <- switch(what, "pda" = c("StTm","EdTm"),
+                         "chromatogram" = c("ACQ$StartTm#1","ACQ$EndTm#1"))
+  start_time <- try(get_metadata_field(sz_method, fields[1])/60000, silent = TRUE)
+  end_time <- try(get_metadata_field(sz_method, fields[2])/60000, silent = TRUE)
+  if (inherits(start_time, "numeric") & inherits(end_time, "numeric")){
+    seq(from = start_time, to = end_time, length.out = nval)
+  } else NA
 }
 
 #' Get number of rows from LCD 3D Data item
@@ -145,7 +298,7 @@ get_shimadzu_rows <- function(path){
 #' Read 'Shimadzu' LCD 3D Raw Data
 #' @author Ethan Bass
 #' @noRd
-read_shimadzu_raw <- function(path, n_lambdas = NULL){
+read_shimadzu_pda <- function(path, n_lambdas = NULL){
   path_raw <- export_stream(path, stream =  c("PDA 3D Raw Data", "3D Raw Data"),
                             verbose = TRUE)
   f <- file(path_raw, "rb")
@@ -172,6 +325,31 @@ read_shimadzu_raw <- function(path, n_lambdas = NULL){
     mat <- mat[-which(is.na(mat[,1])),]
   }
   mat
+}
+
+#' Check OLE stream size
+#' @noRd
+check_streams <- function(path, what = c("pda","chromatogram"), boolean=FALSE){
+  what <- match.arg(what, c("pda","chromatogram"))
+  trace_file <- reticulate::import("olefile")
+  ole <- olefile$OleFileIO(path)
+  if (what == "pda"){
+    pda_exists <- ole$get_size("PDA 3D Raw Data/3D Raw Data") > 0
+    if (boolean){
+      return(pda_exists)
+    } else if (!pda_exists){
+      stop("PDA stream could not be found.")
+    }
+  } else {
+    streams <- ole$listdir()
+    chrom_streams <- streams[grep("Chromatogram Ch", streams)]
+    sizes <- sapply(chrom_streams, function(x) ole$get_size(paste0(x, collapse="/")))
+    if (boolean){
+      return(any(sizes > 0))
+    } else{
+    chrom_streams[which(sizes > 0)]
+    }
+  }
 }
 
 #' Export OLE stream
@@ -220,7 +398,7 @@ read_shimadzu_wavelengths <- function(path){
   count <- 1
   # lambdas <- numeric(n_lambda)
   lambdas <- sapply(seq_len(n_lambda), function(i){
-    readBin(f, what="integer", size = 4)/100
+    readBin(f, what = "integer", size = 4)/100
   })
   lambdas
 }
@@ -245,26 +423,30 @@ decode_shimadzu_block <- function(file) {
   count <- 1
   buffer <- list(0,0,0,0)
 
-  for (i in c(1:2)){
+  while (count < length(signal)){
     n_bytes <- readBin(file, "integer", n = 1, size = 2)
     start <- seek(file, NA, "current")
-
+    if (length(n_bytes) == 0) break
     while(seek(file, NA, "current") < start + n_bytes){
       buffer[[3]] <- readBin(file, "raw", n = 1, size = 1)
-      hex1 <- as.numeric(substr(buffer[[3]], 1, 1))
-      if (hex1 == 0){
-        buffer[[2]] <- strtoi(buffer[[3]], 16)
+      hex1 <- as.numeric(substr(buffer[[3]], start = 1, stop = 1))
+      if (buffer[[3]] == "82"){
+        next
+      } else if (buffer[[3]] == "00"){
+        buffer[[2]] <- 0
+      } else if (hex1 == 0){
+        buffer[[2]] <- strtoi(buffer[[3]], base = 16)
       } else if (hex1 == 1){
-        bin <- as_binary(strtoi(buffer[[3]], 16), 8)
-        buffer[[2]] <- twos_complement(substr(bin, 5, nchar(bin)))
+        bin <- as_binary(strtoi(buffer[[3]], base = 16), n = 8)
+        buffer[[2]] <- twos_complement(substr(bin, start = 5, stop = nchar(bin)))
       } else if (hex1 > 1){
         buffer[[4]] <- readBin(file, "raw", n = (hex1 %/% 2), size = 1)
-        bin <- paste(as_binary(strtoi(c(buffer[[3]], buffer[[4]]), 16), 8),
+        bin <- paste(as_binary(strtoi(c(buffer[[3]], buffer[[4]]), base = 16), n = 8),
                      collapse = "")
         if (hex1 %% 2 == 0){
-          buffer[[2]] <- strtoi(substr(bin, 5, nchar(bin)), 2)
+          buffer[[2]] <- strtoi(substr(bin, start = 5, stop = nchar(bin)), base = 2)
         } else {
-          buffer[[2]] <- twos_complement(substr(bin, 5, nchar(bin)))
+          buffer[[2]] <- twos_complement(substr(bin, start = 5, stop = nchar(bin)))
         }
       }
       buffer[[1]] <- buffer[[1]] + buffer[[2]]