+++ R/zzz.R
@@ -5,4 +5,5 @@
   try({csv <<- reticulate::import("csv", delay_load = TRUE)}, silent = TRUE)
   try({rb_read <<- reticulate::import("rainbow.__init__", delay_load=TRUE)}, silent = TRUE)
   try({rb_parse_agilent <<- reticulate::import("rainbow.agilent", delay_load=TRUE)}, silent = TRUE)
+  try({olefile <<- reticulate::import("olefile", delay_load=TRUE)}, silent = TRUE)
 }