+++ tests/testthat/helpers.R
@@ -3,7 +3,7 @@
 elementwise.all.equal <- Vectorize(function(x, y, ...) {isTRUE(all.equal(x, y, ...))})
 
 # helper function to skip tests if we don't have the right python dependencies
-skip_if_missing_dependecies <- function(reqs = c("scipy","numpy", "aston", "pandas", "olefile")) {
+skip_if_missing_dependencies <- function(reqs = c("scipy","numpy", "aston", "pandas", "olefile")) {
   have_reqs <- sapply(reqs, reticulate::py_module_available)
   if (mean(have_reqs) < 1)
     skip(paste("required packages", reqs[!have_reqs],