+++ tests/testthat/test-read_chroms.R
@@ -6,7 +6,7 @@ path_uv <- test_path("testdata/dad1.uv") #chemstation 131
 x <- read_chroms(path_csv, format_in = "chemstation_csv", progress_bar = FALSE)
 
 test_that("aston parser can read `Agilent Chemstation` 131 files", {
-  skip_if_missing_dependecies()
+  skip_if_missing_dependencies()
   paths <- rep(path_uv, 2)
   x1 <- read_chroms(paths, format_in = "chemstation_uv", parser = "aston",
                     find_files = FALSE,
@@ -94,7 +94,7 @@ test_that("get_filetype works as expected", {
 })
 
 test_that("Rainbow parser can read chemstation 131 files", {
-  skip_if_missing_dependecies()
+  skip_if_missing_dependencies()
   skip_on_cran()
   skip_on_ci()
 
@@ -122,7 +122,7 @@ test_that("Rainbow parser can read chemstation 131 files", {
 })
 
 test_that("chromconverter parser can read chemstation 130 files", {
-  skip_if_missing_dependecies()
+  skip_if_missing_dependencies()
   skip_on_cran()
   x1 <- read_chroms(test_path("testdata/chemstation_130.ch"), progress_bar = FALSE)
   # expect_equal(as.numeric(x[[1]][,1]), as.numeric(x1[[1]][,"220"]))